backend/requirements.txt
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
Markdown==3.4.4
psycopg==3.1.10
psycopg-binary==3.1.10